+++ week04_05/Leeyeeun/everytime/src/main/java/org/skhu/everytime/community/board/api/controller/BoardController.java
@@ -6,7 +6,7 @@ import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import lombok.RequiredArgsConstructor;
-import org.skhu.everytime.community.board.api.dto.BoardResponseDto;
+import org.skhu.everytime.community.board.api.dto.response.BoardResponseDto;
 import org.skhu.everytime.community.board.application.BoardService;
 import org.skhu.everytime.community.board.application.PostService;
 import org.skhu.everytime.community.board.domain.Board;
